requirements.txt
python-dotenv
django-mailjet